requirements.txt
flask
flask-cors
PyPDF2
numpy
requests
python-dotenv